#869 Separate general abstract class PropertyWindow from GraphElementPropertyWindow. PropertyWindow is reusable.
